-Implemented new render pipeline using AZDO OpenGL features (Bindless Textures, MultiDrawElementsIndirect, PMB) -Started first pass on adapting derived Hitable classes to new render pipeline (Primitive, Bumper and Surface so far) -Add new shader based on basicShader to support new pipeline -Updated GLAD to beta 2.0.0 -Added Tracy support for performance profiling when debugging (client download required, use RELEASE_SDL_TRACY build target)

+++ meshes/bumperBaseMesh.h
@@ -521,7 +521,8 @@ const Vertex3D_NoTex2 bumperBase[517]=
 { -0.198152f,0.196647f,-0.003978f, 0.928000f,-0.328000f,-0.176600f, 0.478323f,0.185920f}
 };
 
-const WORD bumperBaseIndices[2352]=
+//const WORD bumperBaseIndices[2352]=
+const unsigned int bumperBaseIndices[2352]=
 {
 	0, 1, 2,
 	0, 2, 3,
